# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Car Rental API
  version: "1.0"
servers:
  - url: http://localhost:5000 # Replace with your actual server URL
    description: Local development server
paths:
  /register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                # Add other properties here
      responses:
        "201":
          description: User registered successfully
        # Add other responses here
  /login:
    post:
      summary: Log in and obtain an access token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  # Add other properties here
...